Change items order in Settings page
