+++ app/src/main/java/io/horizontalsystems/bankwallet/modules/contact/ContactPresenter.kt
@@ -21,10 +21,6 @@ class ContactPresenter(
         router.openTelegram(interactor.walletHelpTelegramGroup)
     }
 
-    override fun didTapAppStatus() {
-        router.openAppStatus()
-    }
-
     override fun didFailSendMail() {
         interactor.copyToClipboard(interactor.email)
         view.showCopied()
