+++ app/src/main/java/io/horizontalsystems/bankwallet/modules/settings/main/MainSettingsPresenter.kt
@@ -27,6 +27,10 @@ class MainSettingsPresenter(
         router.showManageKeys()
     }
 
+    override fun didTapAppStatus() {
+        router.openAppStatus()
+    }
+
     override fun didTapSecurity() {
         router.showSecuritySettings()
     }
