+++ app/src/main/res/layout/fragment_settings.xml
@@ -68,15 +68,15 @@
                 app:title="@string/Settings_SecurityCenter" />
 
             <io.horizontalsystems.views.SettingsView
-                android:id="@+id/experimentalFeatures"
+                android:id="@+id/appStatus"
                 android:layout_width="match_parent"
                 android:layout_height="wrap_content"
                 app:bottomBorder="true"
-                app:icon="@drawable/ic_experimental"
+                app:icon="@drawable/ic_app_status"
                 app:layout_constraintEnd_toEndOf="parent"
                 app:layout_constraintStart_toStartOf="parent"
                 app:layout_constraintTop_toBottomOf="@+id/privacySettings"
-                app:title="@string/Settings_ExperimentalFeatures" />
+                app:title="@string/Settings_AppStatus" />
 
             <io.horizontalsystems.views.SettingsView
                 android:id="@+id/notifications"
@@ -86,7 +86,7 @@
                 app:icon="@drawable/ic_notification"
                 app:layout_constraintEnd_toEndOf="parent"
                 app:layout_constraintStart_toStartOf="parent"
-                app:layout_constraintTop_toBottomOf="@+id/experimentalFeatures"
+                app:layout_constraintTop_toBottomOf="@+id/appStatus"
                 app:title="@string/Settings_Notifications" />
 
             <io.horizontalsystems.views.SettingsView
@@ -113,7 +113,6 @@
                 android:id="@+id/lightMode"
                 android:layout_width="match_parent"
                 android:layout_height="wrap_content"
-                app:bottomBorder="true"
                 app:icon="@drawable/ic_light_mode"
                 app:layout_constraintEnd_toEndOf="parent"
                 app:layout_constraintStart_toStartOf="parent"
@@ -121,15 +120,26 @@
                 app:title="@string/Settings_LightMode" />
 
             <io.horizontalsystems.views.SettingsView
-                android:id="@+id/about"
+                android:id="@+id/experimentalFeatures"
                 android:layout_width="match_parent"
                 android:layout_height="wrap_content"
-                android:layout_marginTop="32dp"
-                app:icon="@drawable/ic_attention"
+                app:bottomBorder="true"
+                app:icon="@drawable/ic_experimental"
                 app:layout_constraintEnd_toEndOf="parent"
                 app:layout_constraintStart_toStartOf="parent"
                 app:layout_constraintTop_toBottomOf="@+id/lightMode"
-                app:title="@string/Settings_About" />
+                app:title="@string/Settings_ExperimentalFeatures" />
+
+            <io.horizontalsystems.views.SettingsView
+                android:id="@+id/contact"
+                android:layout_width="match_parent"
+                android:layout_height="wrap_content"
+                android:layout_marginTop="32dp"
+                app:icon="@drawable/ic_report"
+                app:layout_constraintEnd_toEndOf="parent"
+                app:layout_constraintStart_toStartOf="parent"
+                app:layout_constraintTop_toBottomOf="@+id/experimentalFeatures"
+                app:title="@string/Settings_Report" />
 
             <io.horizontalsystems.views.SettingsView
                 android:id="@+id/shareApp"
@@ -138,19 +148,19 @@
                 app:icon="@drawable/ic_share"
                 app:layout_constraintEnd_toEndOf="parent"
                 app:layout_constraintStart_toStartOf="parent"
-                app:layout_constraintTop_toBottomOf="@+id/about"
+                app:layout_constraintTop_toBottomOf="@+id/contact"
                 app:title="@string/Settings_ShareThisWallet" />
 
             <io.horizontalsystems.views.SettingsView
-                android:id="@+id/report"
+                android:id="@+id/terms"
                 android:layout_width="match_parent"
                 android:layout_height="wrap_content"
+                app:icon="@drawable/ic_terms"
                 app:bottomBorder="true"
-                app:icon="@drawable/ic_report"
                 app:layout_constraintEnd_toEndOf="parent"
                 app:layout_constraintStart_toStartOf="parent"
                 app:layout_constraintTop_toBottomOf="@+id/shareApp"
-                app:title="@string/Settings_Report" />
+                app:title="@string/Settings_Terms" />
 
             <TextView
                 android:id="@+id/appName"
@@ -162,7 +172,7 @@
                 android:textColor="@color/grey"
                 app:layout_constraintEnd_toEndOf="parent"
                 app:layout_constraintStart_toStartOf="parent"
-                app:layout_constraintTop_toBottomOf="@id/report"
+                app:layout_constraintTop_toBottomOf="@id/terms"
                 tools:text="Wallet Bank 1.01" />
 
             <View
