+++ components/icons/src/main/res/drawable/ic_terms.xml
@@ -1,0 +1,10 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="24dp"
+    android:height="24dp"
+    android:viewportWidth="24"
+    android:viewportHeight="24">
+    <path
+        android:fillColor="#000"
+        android:fillType="evenOdd"
+        android:pathData="M18 2c2.21 0 4 1.79 4 4v12c0 2.21-1.79 4-4 4H6c-2.21 0-4-1.79-4-4V6c0-2.21 1.79-4 4-4h12zM7 6C6.735 6 6.48 6.105 6.293 6.293 6.105 6.48 6 6.735 6 7c0 0.265 0.105 0.52 0.293 0.707C6.48 7.895 6.735 8 7 8h10c0.265 0 0.52-0.105 0.707-0.293C17.895 7.52 18 7.265 18 7c0-0.265-0.105-0.52-0.293-0.707C17.52 6.105 17.265 6 17 6H7zm0 5c-0.265 0-0.52 0.105-0.707 0.293C6.105 11.48 6 11.735 6 12c0 0.265 0.105 0.52 0.293 0.707C6.48 12.895 6.735 13 7 13h10c0.265 0 0.52-0.105 0.707-0.293C17.895 12.52 18 12.265 18 12c0-0.265-0.105-0.52-0.293-0.707C17.52 11.105 17.265 11 17 11H7zm-0.707 5.293C6.48 16.105 6.735 16 7 16h4c0.265 0 0.52 0.105 0.707 0.293C11.895 16.48 12 16.735 12 17c0 0.265-0.105 0.52-0.293 0.707C11.52 17.895 11.265 18 11 18H7c-0.265 0-0.52-0.105-0.707-0.293C6.105 17.52 6 17.265 6 17c0-0.265 0.105-0.52 0.293-0.707z" />
+</vector>
